feat(validation): validate openId
