+++ package.json
@@ -56,7 +56,8 @@
     "ts-node": "^8.10.1",
     "tslint": "^5.20.1",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "^3.7.3"
+    "typescript": "^3.7.3",
+    "@types/hapi__joi": "^17.1.0"
   },
   "config": {
     "commitizen": {
@@ -64,6 +65,7 @@
     }
   },
   "dependencies": {
+    "@hapi/joi": "^17.1.1",
     "axios": "^0.18.0",
     "deepmerge": "^4.2.2",
     "express": "^4.17.1",
