+++ src/auth/oidc/OpenIDMetadata.ts
@@ -1,0 +1,12 @@
+import { ClientMetadata } from 'openid-client'
+
+export interface OpenIDMetadata extends ClientMetadata {
+    discovery_endpoint: string
+    issuer_url: string
+    prompt?: 'login'
+    redirect_uri: string
+    scope: string
+    sessionKey?: string
+    isAuthRouteName?: string
+    logout_url: string
+}
